fix warning
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@9101 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/spatialindex/storagemanager/RandomEvictionsBuffer.cc
@@ -63,7 +63,7 @@ IBuffer* SpatialIndex::StorageManager::createNewRandomEvictionsBuffer(IStorageMa
 
 RandomEvictionsBuffer::RandomEvictionsBuffer(IStorageManager& sm, Tools::PropertySet& ps) : Buffer(sm, ps)
 {
-	srand48(time(NULL));
+	srand48((long)time(NULL));
 }
 
 RandomEvictionsBuffer::~RandomEvictionsBuffer()
